fix(zk): proof compatiblity between 32/64b platforms

+++ tfhe-zk-pok/src/proofs/pke.rs
@@ -295,7 +295,7 @@ pub fn prove<G: Curve>(
 
     let x_bytes = &*[
         q.to_le_bytes().as_slice(),
-        d.to_le_bytes().as_slice(),
+        (d as u64).to_le_bytes().as_slice(),
         b_i.to_le_bytes().as_slice(),
         t.to_le_bytes().as_slice(),
         &*a.iter().flat_map(|&x| x.to_le_bytes()).collect::<Box<_>>(),
@@ -684,7 +684,7 @@ pub fn verify<G: Curve>(
 
     let x_bytes = &*[
         q.to_le_bytes().as_slice(),
-        d.to_le_bytes().as_slice(),
+        (d as u64).to_le_bytes().as_slice(),
         b_i.to_le_bytes().as_slice(),
         t.to_le_bytes().as_slice(),
         &*a.iter().flat_map(|&x| x.to_le_bytes()).collect::<Box<_>>(),
