+++ tfhe-zk-pok/src/proofs/range.rs
@@ -190,7 +190,7 @@ pub fn prove<G: Curve>(
         G::Zp::hash(
             core::slice::from_mut(s),
             &[
-                &i.to_le_bytes(),
+                &(i as u64).to_le_bytes(),
                 v_hat.to_bytes().as_ref(),
                 c_hat.to_bytes().as_ref(),
                 c_y.to_bytes().as_ref(),
@@ -286,7 +286,7 @@ pub fn verify<G: Curve>(
         G::Zp::hash(
             core::slice::from_mut(s),
             &[
-                &i.to_le_bytes(),
+                &(i as u64).to_le_bytes(),
                 v_hat.to_bytes().as_ref(),
                 c_hat.to_bytes().as_ref(),
                 c_y.to_bytes().as_ref(),
